+++ CHANGELOG.md
@@ -7,8 +7,9 @@
 * new widget: ToolWidget ("tools") showing tools currently connected to input devices
 * improved search widget: now highlights search results
 * different installation path for translation files
-* copy paste to/from SVG: useable now, but paste only as picture (no erasing possible).
+* copy and paste to/from image formats:
     * this adds the Qt SVG module as a dependency
+    * paste only as picture (no erasing possible)
 * bug fixes
 * internal: different drawing history, memory management, and stacking order mechanism